fix: exponer errores reales del backend y mejorar manejo JWT
- SecurityConfig: agregar /error a permitAll para que Tomcat pueda
  re-despachar errores sin que Spring Security los bloquee con 401
- application.properties: server.error.include-message=always para
  que el campo message llegue al frontend en respuestas de error
- JwtAuthFilter: capturar JwtException (token malformado) y loggear
  JWTs inválidos/expirados en lugar de lanzar excepción
- VentaService: mostrar cantidades de stock como enteros (intValue)
  en lugar de decimales (ej. "Disponible: 3" en vez de "3.0")

+++ backend/src/main/java/com/taller/sae/service/VentaService.java
@@ -113,8 +113,8 @@ public class VentaService {
                                     "Producto no encontrado: " + det.productoId()));
                     throw new ResponseStatusException(HttpStatus.BAD_REQUEST,
                             "Stock insuficiente para: " + p.getNombre()
-                                    + ". Disponible: " + stock.getCantidadActual()
-                                    + ", Requerido: " + det.cantidad());
+                                    + ". Disponible: " + stock.getCantidadActual().intValue()
+                                    + ", Requerido: " + det.cantidad().intValue());
                 }
             }
         }
